backend/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-dateutil==2.9.0.post0
xlsxwriter==3.2.2
reportlab==4.3.1

